+++ tests/integration/ha/helpers.py
@@ -298,7 +298,7 @@ def get_super_password(ops_test: OpsTest, app_name: str = APP_NAME) -> str:
 
 
 async def send_control_signal(
-    ops_test: OpsTest, unit_name: str, kill_code: str, app_name: str = APP_NAME
+    ops_test: OpsTest, unit_name: str, signal: str, app_name: str = APP_NAME
 ) -> None:
     """Issues given job control signals to a ZooKeeper process on a given Juju unit.
 
@@ -313,12 +313,12 @@ async def send_control_signal(
         await ops_test.model.applications[app_name].add_unit(count=1)
         await ops_test.model.wait_for_idle(apps=[app_name], status="active", timeout=1000)
 
-    kill_cmd = f"exec --unit {unit_name} -- pkill --signal {kill_code} -f {PROCESS}"
-    return_code, _, _ = await ops_test.juju(*kill_cmd.split())
+    kill_cmd = f"exec --unit {unit_name} -- pkill --signal {signal} -f {PROCESS}"
+    return_code, stdout, stderr = await ops_test.juju(*kill_cmd.split())
 
     if return_code != 0:
         raise Exception(
-            f"Expected kill command {kill_cmd} to succeed instead it failed: {return_code}"
+            f"Expected kill command {kill_cmd} to succeed instead it failed: {return_code}, {stdout}, {stderr}"
         )
