+++ tests/integration/ha/test_ha.py
@@ -19,6 +19,16 @@ APP_NAME = METADATA["name"]
 USERNAME = "super"
 
 CLIENT_TIMEOUT = 10
+RESTART_DELAY = 60
+
+
+@pytest.fixture()
+async def restart_delay(ops_test: OpsTest):
+    for unit in ops_test.model.applications[APP_NAME].units:
+        await helpers.patch_restart_delay(ops_test=ops_test, unit_name=unit.name, delay=60)
+    yield
+    for unit in ops_test.model.applications[APP_NAME].units:
+        await helpers.remove_restart_delay(ops_test=ops_test, unit_name=unit.name)
 
 
 @pytest.mark.abort_on_fail
@@ -31,6 +41,21 @@ async def test_deploy_active(ops_test: OpsTest):
         storage={"data": {"pool": "lxd-btrfs", "size": 10240}},
     )
     await helpers.wait_idle(ops_test)
+
+
+async def test_replication(ops_test: OpsTest):
+    host_0 = ops_test.model.applications[APP_NAME].units[0].public_address
+    host_1 = ops_test.model.applications[APP_NAME].units[1].public_address
+    host_2 = ops_test.model.applications[APP_NAME].units[2].public_address
+    model_full_name = ops_test.model_full_name
+    password = helpers.get_password(model_full_name or "")
+
+    helpers.write_key(host=host_0, password=password)
+    await asyncio.sleep(1)
+
+    helpers.check_key(host=host_0, password=password)
+    helpers.check_key(host=host_1, password=password)
+    helpers.check_key(host=host_2, password=password)
 
 
 @pytest.mark.abort_on_fail
@@ -227,6 +252,57 @@ async def test_network_cut_self_heal(ops_test: OpsTest, request):
     )
     assert last_write == last_write_leader
     assert total_writes == total_writes_leader
+
+
+async def test_full_cluster_crash(ops_test: OpsTest, request, restart_delay):
+    hosts = helpers.get_hosts(ops_test)
+    leader_name = helpers.get_leader_name(ops_test, hosts)
+    leader_host = helpers.get_unit_host(ops_test, leader_name)
+    password = helpers.get_super_password(ops_test)
+    parent = request.node.name
+    non_leader_hosts = ",".join([host for host in hosts.split(",") if host != leader_host])
+
+    logger.info("Starting continuous_writes...")
+    cw.start_continuous_writes(parent=parent, hosts=hosts, username=USERNAME, password=password)
+    await asyncio.sleep(10)
+
+    logger.info("Counting writes are running at all...")
+    assert cw.count_znodes(parent=parent, hosts=hosts, username=USERNAME, password=password)
+
+    # kill all units "simultaneously"
+    await asyncio.gather(
+        *[
+            helpers.send_control_signal(ops_test, unit.name, signal="SIGKILL")
+            for unit in ops_test.model.applications[APP_NAME].units
+        ]
+    )
+
+    # Check that all servers are down at the same time
+    assert await helpers.all_db_processes_down(ops_test), "Not all units down at the same time."
+
+    await asyncio.sleep(RESTART_DELAY * 2)
+
+    logger.info("Checking writes are increasing...")
+    writes = cw.count_znodes(
+        parent=parent, hosts=non_leader_hosts, username=USERNAME, password=password
+    )
+    await asyncio.sleep(CLIENT_TIMEOUT * 3)  # letting client set up and start writing
+    new_writes = cw.count_znodes(
+        parent=parent, hosts=non_leader_hosts, username=USERNAME, password=password
+    )
+    assert new_writes > writes, "writes not continuing to ZK"
+
+    logger.info("Stopping continuous_writes...")
+    cw.stop_continuous_writes()
+
+    logger.info("Counting writes on surviving units...")
+    last_write = cw.get_last_znode(
+        parent=parent, hosts=non_leader_hosts, username=USERNAME, password=password
+    )
+    total_writes = cw.count_znodes(
+        parent=parent, hosts=non_leader_hosts, username=USERNAME, password=password
+    )
+    assert last_write == total_writes
 
 
 @pytest.mark.abort_on_fail
