+++ tox.ini
@@ -21,6 +21,11 @@ set_env =
     PYTHONPATH = {tox_root}/lib:{[vars]src_path}
     PYTHONBREAKPOINT=ipdb.set_trace
     PY_COLORS=1
+    provider: TEST_FILE=test_provider.py
+    password-rotation: TEST_FILE=test_password_rotation.py
+    tls: TEST_FILE=test_tls.py
+    upgrade: TEST_FILE=test_upgrade.py
+    ha: TEST_FILE=ha/test_ha.py
 pass_env =
     PYTHONPATH
     CHARM_BUILD_DIR
@@ -74,7 +79,7 @@ commands =
         -m pytest -v --tb native -s {posargs} {[vars]tests_path}/unit
     poetry run coverage report
 
-[testenv:integration]
+[testenv:integration-{provider,password-rotation,tls,upgrade,ha}]
 description = Run integration tests
 pass_env =
     {[testenv]pass_env}
@@ -82,64 +87,4 @@ pass_env =
     CI_PACKED_CHARMS
 commands =
     poetry install --with integration
-    poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/
-
-[testenv:integration-password-rotation]
-description = Run integration tests for password-rotation
-pass_env =
-    {[testenv]pass_env}
-    CI
-    CI_PACKED_CHARMS
-commands =
-    poetry install --with integration
-    poetry run pytest -vv --no-header --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_password_rotation.py
-
-[testenv:integration-provider]
-description = Run integration tests for provider
-pass_env =
-    {[testenv]pass_env}
-    CI
-    CI_PACKED_CHARMS
-commands =
-    poetry install --with integration
-    poetry run pytest -vv --no-header --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_provider.py
-
-[testenv:integration-tls]
-description = Run integration tests for tls
-pass_env =
-    {[testenv]pass_env}
-    CI
-    CI_PACKED_CHARMS
-commands =
-    poetry install --with integration
-    poetry run pytest -vv --no-header --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_tls.py
-
-[testenv:integration-ha]
-description = Run integration tests for ha
-pass_env =
-    {[testenv]pass_env}
-    CI
-    CI_PACKED_CHARMS
-commands =
-    poetry install --with integration
-    poetry run pytest -vv --no-header --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_ha.py
-
-[testenv:integration-replication]
-description = Run integration tests for ha
-pass_env =
-    {[testenv]pass_env}
-    CI
-    CI_PACKED_CHARMS
-commands =
-    poetry install --with integration
-    poetry run pytest -vv --no-header --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_replication.py
-
-[testenv:integration-upgrade]
-description = Run integration tests for upgrade
-pass_env =
-    {[testenv]pass_env}
-    CI
-    CI_PACKED_CHARMS
-commands =
-    poetry install --with integration
-    poetry run pytest -vv --no-header --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_upgrade.py
+    poetry run pytest -vv --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/{env:TEST_FILE}
